IMDgroup/pymatgen/diffusion/neb.py: Fix search shortest diffusion coverage
* src/IMDgroup/pymatgen/diffusion/neb.py (_pair_post_filter):
Rewrite graph scanning algo using BFS instead of DFS.  Also, make sure
that we really use the shortest possible paths.  We use BFS to allow
searching the smallest possible number of segments to reach a given
image.  (This is not useful right now, but opens the possibility to
limit the compound diffusion paths).